feat: Add PDF conversion and fix cache creation bug
To support text extraction from PDFs, the [pdf] option was added to the markitdown library.

Additionally, an error was occurring in the get_markdown_content function because a DocumentConverterResult object was being passed when calling create_cache.
This has been fixed to use the .markdown attribute to pass the str type that the create_cache function requires.